stayCumbrecita-chatbot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.0
greenlet==3.2.3
openai==1.3.7
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0

# Procesamiento de PDFs
PyPDF2==3.0.1
pypdf==3.17.4

# Utilidades adicionales
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0 